dbwrap: add flags DBWRAP_STORE_PERSISTENT
Request storing a record with persistent semantics, the details of the
implementation are up to the backend.

Fwiw, the existing "flags" DBWRAP_[REPLACE|INSERT|MODIFY] are not flags in the
real sense, but rather something like an opcode/enum.

The new values DBWRAP_STORE_PERSISTENT will be real flags, so the backend
implementations have to be careful when processing the flags, eg if passing the
flags to TDB in the backend.

Signed-off-by: Ralph Boehme <slow@samba.org>
Reviewed-by: Martin Schwenke <martin@meltin.net>