dbwrap: add record flags to struct db_record
This is needed to implement persistent record deletion in the case that
a record was stored as persistent, then fetched and is then stored
without persistency.

The additional tdb record header uses CTDB_REC_FLAG_PERSISTENT to denote if the
record was stored as persistent. When fetching a record we remember the
persistency property in record.flags.persistent.

When a subsequent store doesn't request persistency and record.flags.persistent
is set, we delete record from the persistent database.

Signed-off-by: Ralph Boehme <slow@samba.org>
Reviewed-by: Martin Schwenke <martin@meltin.net>