python/samba: add Prometheus metrics exporter
Prometheus[1] monitoring and alerting system is de facto standard for
storing and querying metrics, in particular within Kubernetes clusters.
Allow Samba to export various internal counters as Prometheus metrics
using existing samba tools (smbstatus). The actual exporter is
implemented in Python and may be executed as a side-car (container) to
the main smbd process. This design decouples the 'what' from the 'how'
and transfers the burden of representing Samba's metrics as Prometheus
HTTP queries to a dedicated Python library[2].

When the '--profile' flag is passed to samba-prometheus-exporter and
Samba is compiled with profiling enabled it will also export a set of
profiling metrics.

[1] https://prometheus.io/
[2] https://prometheus.github.io/client_python/

Signed-off-by: Shachar Sharon <ssharon@redhat.com>