smbd: allow link with tcmalloc
TCMalloc[1] is a memory allocator designed as an alternative to the
default lib C allocator. It is useful in cases where the back-end VFS
driver is multi-threaded with extensive dynamic-memory allocations
(such as vfs_ceph* + libcephfs).

Allow users to build and link against tcmalloc (and override default
allocator) by using the '--with-tcmalloc' configuration option.

[1] https://google.github.io/tcmalloc/design.html

Signed-off-by: Shachar Sharon <ssharon@redhat.com>