smbd: allow link against tcmalloc
TCMalloc[1] is a memory allocator designed as an alternative to the
default lib C allocator, optimized for multi-thread applications. This
may be useful in cases where the back-end VFS driver is multi-threaded
and uses extensive dynamic allocations (e.g., vfs_ceph* + libcephfs).

Allow users to build and link against tcmalloc by using the
'--with-tcmalloc' configuration option.

[1] https://google.github.io/tcmalloc/design.html

Signed-off-by: Shachar Sharon <ssharon@redhat.com>