vfs_ceph_rgw: Handling for empty name, allows file creation inside dir
Why we get empty name? Apparatly this was only causing a crash in
librgw, due to empty name for lookup/open call.

For now we just skip this, but need to find out how to deal with this.
We can't doing strlen for every name recevied. Ofcourse we can optimize
it by just checking if 0th or 1st position is '\0' or not, but ideally
we should not need to check filenames or may be we should.

Signed-off-by: Vinit Agnihotri <vagnihot@redhat.com>