vfs_ceph_rgw: Use rgw_create() instead of rgw_mkdir()
When we create a "directory" with rgw_mkdir(), it doesn't create an object,
just an index entry, so no object actually exists(). This is not legit w.r.t S3,
however librgw makes this for NFS to work better.

For some reason, even if rgw_mkdir() creates a dir, it no longer stays
accessible for samba.

As a trial we shall create dir using rgw_create(), which actually
creates an object in bucket. This way we can open and list contents of
created dirs.

we are still evaluating if this is right way or not.

Signed-off-by: Vinit Agnihotri <vagnihot@redhat.com>
(cherry picked from commit 2e6aea9e34f6c6bbaeb0a71bde09c53de4b7cb21)