scavenger: clean up stale delete-on-close tokens in the scavenger
Currently we discard Durable Handles when processing a client disconnect that
have delete-on-close set, which will then also clean up the delete-on-close
token.

This is going to change for Persistent Handles, so we have to prepare for
cleaning up delete-on-close tokens in the scavenger.

Windows takes it one step further and the file gets actually deleted: on Windows
delete-on-close in set at the FSA layer so when the scavenger closes the handle,
the filesystem will just follow standard delete-on-close processing and delete
the file.

For Samba this is hard to implement in the scavenger, as it would need to be
extended to be able talk to the Samba VFS to delete the file, so we merely
remove the stale delete-token and without deleting the file.

Signed-off-by: Ralph Boehme <slow@samba.org>